requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9